slice via u32s, removing a million casts

+++ win32/src/pe/exports.rs
@@ -26,7 +26,7 @@ unsafe impl x86::Pod for IMAGE_EXPORT_DIRECTORY {}
 impl IMAGE_EXPORT_DIRECTORY {
     #[allow(dead_code)]
     pub fn name<'a>(&self, image: &'a Mem) -> &'a str {
-        image.slice(self.Name as usize..).read_strz()
+        image.slice(self.Name..).read_strz()
     }
 
     pub fn fns<'a>(&self, image: &'a Mem) -> &'a [u32] {
@@ -43,9 +43,7 @@ impl IMAGE_EXPORT_DIRECTORY {
             self.NumberOfNames as usize,
         );
 
-        let ni = names
-            .iter()
-            .map(|&addr| image.slice(addr as usize..).read_strz());
+        let ni = names.iter().map(|&addr| image.slice(addr..).read_strz());
         let oi = ords.iter().copied();
         ni.zip(oi)
     }
@@ -56,6 +54,6 @@ pub fn read_exports<'a>(
     base: u32,
     exports: &'a IMAGE_DATA_DIRECTORY,
 ) -> &'a IMAGE_EXPORT_DIRECTORY {
-    let image = &machine.x86.mem.slice(base as usize..);
+    let image = &machine.x86.mem.slice(base..);
     exports.as_mem(image).view::<IMAGE_EXPORT_DIRECTORY>(0)
 }
