+++ win32/src/pe/file.rs
@@ -92,9 +92,7 @@ pub struct IMAGE_DATA_DIRECTORY {
 unsafe impl x86::Pod for IMAGE_DATA_DIRECTORY {}
 impl IMAGE_DATA_DIRECTORY {
     pub fn as_mem<'a>(&self, image: &'a Mem) -> &'a Mem {
-        &image
-            .slice(self.VirtualAddress as usize..)
-            .slice(..self.Size as usize)
+        &image.slice(self.VirtualAddress..).slice(..self.Size)
     }
 }
